enrichment_worker/requirements.txt
psycopg2-binary
google-api-python-client
youtube-transcript-api
